Nostr Identity Implementation
This contains: the ONLY on every initial page load it will check for the NIP-07 ID, this is embeded in the Header using the Nostr Indenty Block.

Some Notes:
I can confirm this works using Chrome &  Nos2x, but can't confirm it working with Alby.

Performance: this thing stutters like no tomorrow, will see how this works on prod.
But so far not impressed 😅